This version has a working simulator, and a synthesized version to be tested on Linux.
There is SPI send and receive, but they just talk to each other.  One
step at a time, after this receive will be tested with the spi send
from the desktop using the pico-ice sdk.

Surprisingly it only uses 24 BRAMs even though there are 8K words.  How is that possible?

Using yosys -dsp, to shrink the number of luts from 2.7K to about 2.1K

A lot of the top level stuff was moved to shared.v, so that it can be used in
both simulation and synthesis.

The timer was removed, but the code was kept in common-verliog in case anyone
needs it.  Easy to find that way.

There were a couple of values which were reg or wire, and had to
be switched to work in synthesis.